requirements.txt
openai-agents>=0.0.9
streamlit>=1.43.0
pandas>=2.2.0
plotly>=5.22.0
python-dotenv>=1.0.0
tabulate
